v0.1.6: Add `--drop-row-if-null=COL` flag (fixes #9)
We now support a `--drop-row-if-null=COL` flag which can be used to
automatically discard certain lines when particular columns are NULL.
This argument can be passed more than once to check more than one
column, and it is applied _after_ will normalize NULLs as specified by
`--null`.

In order to implement this, we had to switch from `docopt` (a deprecated
argument parser) to `structopt` (the recommended replacement).